+++ both/views/Contact/ContactSummary.jsx
@@ -6,22 +6,23 @@ ContactSummary = React.createClass({
 	},
 
 	render() {
-	    var contact = this.props.item;
+		var contact, profile;
+	    contact = this.props.item;
 	    if(!contact) {
 	    	return <div/>
 	    }
 	    if(contact.getProfile) {
-	    	contact = contact.getProfile();
+	    	profile = contact.getProfile();
 	    }
-	    if(contact.name) {
-			this.seed = contact.name.charCodeAt(0)+contact.name.charCodeAt(1)+contact.name.charCodeAt(2);
+	    if(profile.name) {
+			this.seed = profile.name.charCodeAt(0)+profile.name.charCodeAt(1)+profile.name.charCodeAt(2);
 		}
 		else {
 			this.seed = 1;
 		}
 	    var cardStyle = Math.ceil(this.random()*3);
 	    var size=this.props.size;
-	    var services = contact&&contact.services?contact.services.join(' | '):'';
+	    var services = profile&&profile.services?profile.services.join(' | '):'';
 	    return (
 	    	<div className={"business-card"+" business-card-style-"+cardStyle+" "+size}>
 	    		
@@ -31,13 +32,13 @@ ContactSummary = React.createClass({
 					<input type="checkbox" />
 				</div>*/}
 				<div className="contact-thumbnail pull-left">
-				    <img alt="image" src={contact.thumb} />
+				    <img alt="image" src={contact.getThumbUrl()} />
 				 </div>
 				 <div className="contact-info">
 				 	<div>
-						<h2>{contact.name}</h2>
-						<b>Email</b> {contact.email}<br/>
-						<b>Phone</b> {contact.phone}<br/>
+						<h2>{contact.getName()}</h2>
+						<b>Email</b> {profile.email}<br/>
+						<b>Phone</b> {profile.phone}<br/>
 						<div style={{margin:"10px 0 10px 70px",borderBottom:"1px solid #ccc"}}>
 						</div>
 						{services}