+++ server/publish.jsx
@@ -24,8 +24,8 @@ Meteor.publish('teamsAndFacilitiesForUser', function () {
 	teams.forEach(function(t){
 		teamIds.push(t._id);
 	});
-	facilities = Facilities.find({"_team._id":{$in:teamIds}});
-	issues = Issues.find({"_team._id":{$in:teamIds}},{sort: {createdAt: -1}});
+	facilities = Facilities.find({"team._id":{$in:teamIds}});
+	issues = Issues.find({"team._id":{$in:teamIds}},{sort: {createdAt: -1}});
 	return [teams,facilities,issues];
 });